backend/requirements.txt
# =============================================================================
# CONEIC 2027 - Python Dependencies
# =============================================================================

# Django core
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Cache & message broker
redis>=5.0,<6.0

# Async task queue
celery[redis]>=5.4,<6.0

# Email
django-anymail[resend]>=11.0,<13.0

# Security
argon2-cffi>=23.1,<24.0

# Environment & config
python-decouple>=3.8,<4.0

# Static files
whitenoise>=6.6,<7.0

# Image processing
Pillow>=10.2,<11.0

# QR code generation
qrcode[pil]>=7.4,<8.0

# PDF generation
reportlab>=4.1,<5.0

# Excel export
openpyxl>=3.1,<4.0

# HTTP client
requests>=2.31,<3.0

# Production WSGI server
gunicorn>=22.0,<23.0
gevent>=24.2,<25.0
